Separo JavaScript en pseudo-MVC
