+++ js/model/ode-solver.js
@@ -1,4 +1,4 @@
-class EulerODESolver {
+class ODESolver {
     constructor(model) {
         this.size = model.size;
         this.initialConditions = model.initialConditions;
@@ -34,26 +34,4 @@ class EulerODESolver {
     }
 }
 
-class SIRModel {
-    constructor(susceptible, infected, recovered, transmissionRate, recoveryRate) {
-        this.size = 3;
-
-        this.totalPopulation = susceptible + infected + recovered;
-        this.initialConditions = [susceptible, infected, recovered];
-
-        this.equations = [
-            (currentConditions) => {
-                const [susceptible, infected] = currentConditions;
-                return -transmissionRate * susceptible * infected / this.totalPopulation;
-            },
-            (currentConditions) => {
-                const [susceptible, infected] = currentConditions;
-                return transmissionRate * susceptible * infected / this.totalPopulation - recoveryRate * infected;
-            },
-            (currentConditions) => {
-                const [_, infected] = currentConditions;
-                return recoveryRate * infected;
-            }
-        ];
-    }
-}
+export { ODESolver };
